refactor: separate queue rendering from single-toast presentation
