+++ app/components/ui/toast/HToastList.vue
@@ -1,0 +1,46 @@
+<template>
+  <Snackbar.Portal teleport="body">
+    <Snackbar.Queue
+      v-slot="{ items }"
+      :class="$style.queue"
+    >
+      <TransitionGroup name="snack-list-slide">
+        <HToast
+          v-for="(item, i) in items"
+          :key="item.id"
+          :item="item"
+          :index="i"
+        />
+      </TransitionGroup>
+    </Snackbar.Queue>
+  </Snackbar.Portal>
+</template>
+
+<script setup lang="ts">
+import { Snackbar } from "@vuetify/v0"
+import HToast from "./HToast.vue"
+</script>
+
+<style module>
+.queue {
+  position: fixed;
+  bottom: var(--spacing-2xl);
+  right: var(--spacing-2xl);
+}
+</style>
+
+<style scoped>
+.snack-list-slide-move,
+.snack-list-slide-enter-active,
+.snack-list-slide-leave-active {
+  transition:
+    transform 0.3s ease-out,
+    opacity 0.3s ease-out;
+}
+
+.snack-list-slide-enter-from,
+.snack-list-slide-leave-to {
+  opacity: 0;
+  transform: translateX(70px);
+}
+</style>
